+++ src/kiota/Program.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.CommandLine;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.CommandLine;
 
 namespace kiota;
 static class Program