+++ src/kiota/Rpc/NoopLogger.cs
@@ -1,6 +1,4 @@
-﻿
-using System;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 
 namespace kiota.Rpc;
 public class NoopLogger<T> : ILogger<T>