+++ src/kiota/Rpc/SearchResult.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Kiota.Builder.SearchProviders;
+﻿using Kiota.Builder.SearchProviders;
 
 namespace kiota.Rpc;